Add event functions to events.js
Functions include all of the click handlers and the evaluate function. The click handlers are in a barebones state, all they do at this point is add the player's symbol to the correct index of the gameArray.

The current player is decided by the the combination of the player array and the nextTurn function. nextTurn increments an accumulator, which is used to reference a point in the player array. When the player array is called, nextTurn is modulo'd by 2 to determine whether the current player is player[0] or player[1]

The evaluate function determines the winner of the game. It leverages a 2D array to work across rows, down columns, and diagonally to look for 3 in a row. If found, it reports the winner.

gameArray is set up to pull the values that exist on the HTML. In the future, when a square is clicked, the HTML will be populated at that specific space with the token of the user. With this methodology, regardless of the token used, the evaluate and gameArray elements will be able to work effectively.